Distribute HistoryElement classes to in different files
Zeitwerk doesnt allow it anymore to declare multiple
classes in one file, since it fetches the constant
names from the filename.
