+++ src/api/app/models/history_element/request_accepted.rb
@@ -1,0 +1,15 @@
+module HistoryElement
+  class RequestAccepted < HistoryElement::Request
+    def color
+      'green'
+    end
+
+    def description
+      'Request got accepted'
+    end
+
+    def user_action
+      'accepted request'
+    end
+  end
+end
