+++ src/api/app/models/history_element/request_all_reviews_approved.rb
@@ -1,0 +1,15 @@
+module HistoryElement
+  class RequestAllReviewsApproved < HistoryElement::Request
+    def color
+      'green'
+    end
+
+    def description
+      'Request got reviewed'
+    end
+
+    def user_action
+      'approved review'
+    end
+  end
+end
