+++ src/api/app/models/history_element/request_declined.rb
@@ -1,0 +1,15 @@
+module HistoryElement
+  class RequestDeclined < HistoryElement::Request
+    def color
+      'maroon'
+    end
+
+    def description
+      'Request got declined'
+    end
+
+    def user_action
+      'declined request'
+    end
+  end
+end
