+++ src/api/app/models/history_element/request_deleted.rb
@@ -1,0 +1,15 @@
+module HistoryElement
+  class RequestDeleted < HistoryElement::Request
+    def color
+      'red'
+    end
+
+    def description
+      'Request was deleted'
+    end
+
+    def user_action
+      'deleted request'
+    end
+  end
+end
