+++ src/api/app/models/history_element/request_priority_change.rb
@@ -1,0 +1,15 @@
+module HistoryElement
+  class RequestPriorityChange < HistoryElement::Request
+    def color
+      'black'
+    end
+
+    def description
+      'Request got a new priority: ' + description_extension
+    end
+
+    def user_action
+      "changed priority to #{description_extension}"
+    end
+  end
+end
