+++ src/api/app/models/history_element/request_revoked.rb
@@ -1,0 +1,15 @@
+module HistoryElement
+  class RequestRevoked < HistoryElement::Request
+    def color
+      'orange'
+    end
+
+    def description
+      'Request got revoked'
+    end
+
+    def user_action
+      'revoked request'
+    end
+  end
+end
