+++ src/api/app/models/history_element/review_accepted.rb
@@ -1,0 +1,11 @@
+module HistoryElement
+  class ReviewAccepted < HistoryElement::Review
+    def description
+      'Review got accepted'
+    end
+
+    def user_action
+      'accepted review'
+    end
+  end
+end
