+++ src/api/app/models/history_element/review_assigned.rb
@@ -1,0 +1,11 @@
+module HistoryElement
+  class ReviewAssigned < HistoryElement::Review
+    def description
+      'Review got assigned'
+    end
+
+    def user_action
+      'assigned review'
+    end
+  end
+end
