+++ src/api/app/models/history_element/review_declined.rb
@@ -1,0 +1,11 @@
+module HistoryElement
+  class ReviewDeclined < HistoryElement::Review
+    def description
+      'Review got declined'
+    end
+
+    def user_action
+      'declined review'
+    end
+  end
+end
