+++ src/api/app/models/history_element/review_obsoleted.rb
@@ -1,0 +1,11 @@
+module HistoryElement
+  class ReviewObsoleted < HistoryElement::Review
+    def description
+      'Review got obsoleted'
+    end
+
+    def user_action
+      'obsoleted review'
+    end
+  end
+end
