+++ src/api/app/models/history_element/review_reopened.rb
@@ -1,0 +1,11 @@
+module HistoryElement
+  class ReviewReopened < HistoryElement::Review
+    def description
+      'Review got reopened'
+    end
+
+    def user_action
+      'reopened review'
+    end
+  end
+end
